backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
aiosqlite==0.21.0
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1
openai==1.109.1
httpx==0.28.1
beautifulsoup4==4.14.2
markdownify==1.2.0
sentence-transformers==5.1.1
numpy==2.3.3
sqlite-vec==0.1.6
tenacity==9.1.2
pyyaml==6.0.2
pytest==8.4.2
pytest-asyncio==1.2.0
